* 'master' of https://github.com/strawbot/EmbeddedTimbre:
  updated updates
  changed limit on dumping memory from 255 lines to 65535
  changed limit on dictionary size from 255 entries to 65534.
  bug fixes caught by compiling with Qt IDE
  added dictionary iterator
  Added script reading and parsing of EEPROM files; ensured a file will end with 0xFF in EEPROM; added in hashed dictionary support
  extended Timeout API with the addition of restartTimeout() which restarts the timeout
  improved pdump
  added option to remove float support; little bit of code tidying
  added signed floating numbers on input line support
  support for positive floating point added; can use decimal point to get a float but negative not supported
  exposed prompt printer for making cleaner outputs from machines
  float words for cli
  added 32 bit multiplatrform support using Whole; Long can be 64 bit
  change value transfer for Longs to be tied to size of long

# Conflicts:
#	timeout.h

+++ library.c
@@ -282,7 +282,7 @@ void debugpmsg(PGM_P m)
 
 void dump(void) // ram dump
 {
-	Byte n = (Byte)*_DROP_;
+	Short n = (Short)*_DROP_;
 	Cell memory = *_DROP_;
 	
 	dotindex((Byte)memory);
@@ -324,9 +324,9 @@ void DOT_S(void)  /* -- */
     immediatenames immediatebodies
     constantnames constantbodies
 */
-Byte search_names(Byte *name, PGM_P dictionary) // return name number or 0 if not found
+Short search_names(Byte *name, PGM_P dictionary) // return name number or 0 if not found
 {
-	Byte index = 1;
+	Short index = 1;
 	
 	while(pgm_read_byte(dictionary))
 	{
@@ -340,7 +340,7 @@ Byte search_names(Byte *name, PGM_P dictionary) // return name number or 0 if no
 
 void search_dictionaries(void) // look through dictionaries for word
 { // s -- a \ f
-	Byte index, length = *(Byte *)(_TOP_);
+	Short index, length = *(Byte *)(_TOP_);
 	Byte name[31+1]; //[length+1];
 	
 	memcpy(&name[0], (Byte *)(_TOP_+1), length);
